Abstract Interface for a Component

AbstractComponent is the abstract base class for components. All subclasses must implement fit and density.

Public fields

pi

the class probability of the component in the mixture model

Methods

Public methods

AbstractComponent$pi_density()

AbstractComponent$density()

AbstractComponent$fit()

AbstractComponent$set_pi()

AbstractComponent$Q()

AbstractComponent$to_fixed_component()

AbstractComponent$clone()

Method pi_density()

Calculate the density multiplied with pi at the given x-coordinates.

Usage

AbstractComponent$pi_density(x)

Arguments

x

a vector of points at which the density is evaluated.

Method density()

Abstract. Calculate the density at the given x-coordinates.

Usage

AbstractComponent$density(x)

Arguments

x

a vector of points at which the density is evaluated.

Method fit()

Abstract. Fits the component to the signal rf by maximizing Q. If performed approximately turns the algorithm into a ECM algorithm.

Usage

AbstractComponent$fit(x, rf)

Arguments

x

coordinates of points on the x-axis.

rf

coordinates of function values.

Method set_pi()

Setter for pi.

Usage

AbstractComponent$set_pi(pi)

Arguments

pi

the new value.

Method Q()

Calculates Q between rf and the density for the component as defined in !!!paper-doi!!!.

Usage

AbstractComponent$Q(x, rf)

Arguments

x

vector of x values.

rf

vector of the product between p_y and the responsiblity vector of the given component.

Method to_fixed_component()

Returns a component with a fixed density identical to the density of the current component.

Usage

AbstractComponent$to_fixed_component(x)

Arguments

x

vector of x values.

Method clone()

The objects of this class are cloneable with this method.

Usage

AbstractComponent$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.