Component which models a truncated voigt profile

Component which models a truncated voigt profile. Models a voigt peak in the signal. Truncation is needed to deal with peaks which have significant mass outside the x range.

Super class

spectralem::AbstractComponent -> ComponentTrunctVoigt

Public fields

pos

position of the voigt profile

gwidth

gaussian width of the voigt profile

lwidth

gaussian width of the voigt profile

Methods

Public methods

ComponentTrunctVoigt$new()

ComponentTrunctVoigt$density()

ComponentTrunctVoigt$fit()

ComponentTrunctVoigt$set_params()

ComponentTrunctVoigt$get_amp()

ComponentTrunctVoigt$clone()

Inherited methods

spectralem::AbstractComponent$Q()

spectralem::AbstractComponent$pi_density()

spectralem::AbstractComponent$set_pi()

spectralem::AbstractComponent$to_fixed_component()

Method new()

Initialize the voigt component.

Usage

ComponentTrunctVoigt$new(min_width, possible_peak_positions, pi = 1)

Arguments

min_width

minimal value for lwidth and gwidth

possible_peak_positions

range of possible peak positions

pi

initial value for pi.

Method density()

Calculate the density at the given x-coordinates.

Usage

ComponentTrunctVoigt$density(x)

Arguments

x

a vector of points at which the density is evaluated.

Method fit()

Fits the component to the signal rf by maximizing Q. Ensures monotonicity by comparing with last fit.

Usage

ComponentTrunctVoigt$fit(x, rf)

Arguments

x

coordinates of points on the x-axis.

rf

coordinates of function values.

Method set_params()

Setter for pos, gwidth and lwidth.

Usage

ComponentTrunctVoigt$set_params(pos = NULL, gwidth = NULL, lwidth = NULL)

Arguments

pos

the new pos

gwidth

the new gwidth

lwidth

the new lwidth

Method get_amp()

Gets amplitude of the voigt profile.

Calculate Q and it's gradient.

Guess pos, gwidth, lwidth by fitting a pseudo voigt. To be treated as a blackbox.

Usage

ComponentTrunctVoigt$get_amp()

Method clone()

The objects of this class are cloneable with this method.

Usage

ComponentTrunctVoigt$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.