Quickly fits voigt profiles to a signal.

Requires no starting fit. Most useful to get a fast and decent fit. The algorithm was developed to produce a fast and decent fit that general non linear optimizers such as optim fail to produce.

Usage

spectralem(
  x,
  y,
  max_peaks,
  max_iter,
  print_progress = TRUE,
  add_component_every_iters = 10,
  start_peaks = list(pos = c(), gwidth = c(), lwidth = c()),
  background_model = list(ComponentTrunctLinear$new()),
  placement_strategy = StrategyMaxSmoothLossfield$new()
)

Arguments

x

the x coordinates of the Raman signal

y

the function values of the function to be fit evaluated at the x coordinates

max_peaks

the maximal number of components to fit

max_iter

the maximal number of iterations

print_progress

(Optional) Defines strategy used to place/birth new peak. Instance of class which implements AbstractStrategy.

add_component_every_iters

How many iteration to perform until a new peak is added. For all peaks to spawn max_iter >= (2+max_peaks) * add_component_every_iters needs to be satisfied.

start_peaks

initial peak parameters allows user to pass know peaks. list of vectors

background_model

defines the background model, support BSplines and a linear background. Instance of class which implements AbstractComponent.

placement_strategy

(Optional) Defines strategy used to place/birth new peak. Instance of class which implements AbstractStrategy.